requirements.txt
streamlit>=1.30.0
PyPDF2>=3.0.0
google-generativeai>=0.3.0
pillow>=10.0.0
python-dotenv>=1.0.0
